+++ app/components/omnibox/controllers/area-controller.js
@@ -49,23 +49,24 @@ angular.module('lizard-nxt')
       });
       angular.forEach(promises, function (promise) {
         promise.then(null, null, function (response) {
-          if (response.data && response.layerSlug === "dem/nl") {
-            $scope.box.content[response.layerGroupSlug]
-              .layers["dem/nl"]
+          if (response.data && response.data !== "null") {
+            switch (response.layerSlug) {
+            case "dem/nl":
               // Since the data is not properly formatted in the back
               // we convert it from degrees to meters here
-              .data = RasterService.handleElevationCurve(response.data);
-          } else if (response.data && response.layerSlug === "radar/basic") {
-            $scope.box.content.rain.layers["radar/basic"].data = response.data;
-            _setFilteredRainData();
+              $scope.box.content.elevation.layers["dem/nl"].data
+                = RasterService.handleElevationCurve(response.data);
+              break;
+            case "radar/basic":
+              $scope.box.content.rain.layers["radar/basic"].data
+                = response.data;
+              _setFilteredRainData();
+              break;
+            }
           }
         });
       });
-      // Draw feedback when all promises resolved
-      //$q.all(promises).then(drawFeedback);
     };
-
-
 
     /**
      * Updates area when user moves map.
@@ -77,13 +78,11 @@ angular.module('lizard-nxt')
 
     $scope.$watch('timeState.at', function (n, o) {
       if (n === o) { return true; }
-      // _setFilteredRainData();
       fillArea($scope.mapState.bounds);
     });
 
     $scope.$watch('timeState.aggWindow', function (n, o) {
       if (n === o) { return true; }
-      // _setFilteredRainData();
       fillArea($scope.mapState.bounds);
     });
 
@@ -102,7 +101,6 @@ angular.module('lizard-nxt')
     $scope.countKeys = UtilService.countKeys;
 
     var _resetFilteredRainData = function () {
-      $scope.filteredRainData = undefined;
       $scope.filteredRainDataPerKilometer = undefined;
     };
 
@@ -116,51 +114,26 @@ angular.module('lizard-nxt')
       }
 
       var filteredData = [],
-          current,
           currentTimestamp,
           currentVal,
           aggWindowStart = $scope.timeState.at,
           aggWindowEnd = aggWindowStart + $scope.timeState.aggWindow,
-          // aggWindowStart = $scope.timeState.at + $scope.timeState.aggWindow,
-          // aggWindowEnd = aggWindowStart + 2 * $scope.timeState.aggWindow,
           squareKilometers = UtilService.extent2kilometers($scope.mapState.bounds),
           DECIMAL_RESOLUTION = 100;
-
-      // we offset the per-day aggregation with 8 hourS
-      // if ($scope.timeState.aggWindow === 86400000) {
-      //   console.log("PER-DAY AGG DETECTED");
-      //   aggWindowStart += 28800000;
-      //   aggWindowEnd += 28800000;
-      // }
-
-      console.log("current aggWindow..: " + $scope.timeState.aggWindow);
-      console.log("interval start.....:" + new Date(aggWindowStart));
-      console.log("interval end.......:" + new Date(aggWindowEnd));
 
       for (var i = 0; i < data.length; i++) {
         currentTimestamp = data[i][0];
         currentVal = data[i][1];
-
         if (currentTimestamp > aggWindowStart && currentTimestamp <= aggWindowEnd) {
-
-          console.log("[+] found point within interval:" + new Date(currentTimestamp));
-
-          $scope.filteredRainData = currentVal;
           $scope.filteredRainDataPerKilometer
-            = Math.round(($scope.filteredRainData / squareKilometers) * DECIMAL_RESOLUTION) / DECIMAL_RESOLUTION;
-
-          console.log("--> $scope.filteredRainData =", $scope.filteredRainData);
-          console.log("--> $scope.filteredRainDataPerKM =", $scope.filteredRainDataPerKilometer);
+            = Math.round((currentVal / squareKilometers) * DECIMAL_RESOLUTION) / DECIMAL_RESOLUTION
+            || "0.00";
           return;
-
-        } else {
-           console.log("[-] found point outside interval:" + new Date(currentTimestamp));
         }
       }
 
       // If a value was found for the current timeState.at/aggWindow combo,
       // this line is never reached:
-      console.log("[-] No point found :(");
       _resetFilteredRainData();
     };